Refactor, fix style, add tests
